requirements.txt
Django==5.2.5
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
Pillow==10.3.0
drf-spectacular==0.28.0
openpyxl==3.1.2
python-decouple==3.8
